New method of rugosity calculations added, it is taking more time but will produce lower overall rugosity and will have less extreme rugosity areas; Fix of incorrect camera rotation paste functionality; Very small models(in terms of dimensions) would be scaled up; Camera speed and far plane is now depend on size of loaded model; Added functionality to support more accurate ray-triangle hit position; Now vertex positions are firstly read into double vector instead of float, it is needed for correct shifting of coordinates from global to local system.

+++ SubSystems/FEObjLoader.h
@@ -42,7 +42,7 @@ namespace FocalEngine
 
 	struct FERawOBJData
 	{
-		std::vector<glm::vec3> rawVertexCoordinates;
+		std::vector<glm::dvec3> rawVertexCoordinates;
 		std::vector<glm::vec3> rawVertexColors;
 		std::vector<glm::vec2> rawTextureCoordinates;
 		std::vector<glm::vec3> rawNormalCoordinates;
